requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
abseil-cpp=20211102.0=hc377ac9_0
accelerate=0.23.0=pypi_0
aiohttp=3.8.5=py311h80987f9_0
aiosignal=1.2.0=pyhd3eb1b0_0
anyio=4.0.0=pyhd8ed1ab_0
appnope=0.1.3=pyhd8ed1ab_0
arrow-cpp=11.0.0=hc7aafb3_2
asttokens=2.4.0=pyhd8ed1ab_0
async-timeout=4.0.2=py311hca03da5_0
attrs=22.1.0=py311hca03da5_0
aws-c-common=0.6.8=h80987f9_1
aws-c-event-stream=0.1.6=h313beb8_6
aws-checksums=0.1.11=h80987f9_2
aws-sdk-cpp=1.8.185=ha71a6ea_1
backcall=0.2.0=pyh9f0ad1d_0
backoff=1.11.1=pyhd8ed1ab_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
bitsandbytes=0.41.1=pypi_0
blas=1.0=openblas
boost-cpp=1.82.0=h48ca7d4_1
bottleneck=1.3.5=py311ha0d4635_0
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
brotlipy=0.7.0=py311h80987f9_1002
bzip2=1.0.8=h620ffc9_4
c-ares=1.19.1=h80987f9_0
ca-certificates=2023.7.22=hf0a4a13_0
cachetools=5.3.1=pyhd8ed1ab_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py311h80987f9_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
chromadb=0.3.25=pyhd8ed1ab_0
click=8.1.7=unix_pyh707e725_0
clickhouse-connect=0.5.22=pypi_0
coloredlogs=15.0.1=pyhd8ed1ab_3
comm=0.1.4=pyhd8ed1ab_0
cryptography=41.0.3=py311hd4332d6_0
dataclasses=0.8=pyh6d0b6a4_7
dataclasses-json=0.5.7=pyhd8ed1ab_0
datasets=2.14.5=py_0
debugpy=1.6.7=py311h313beb8_0
decorator=5.1.1=pyhd8ed1ab_0
dill=0.3.7=py311hca03da5_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
faiss=1.7.4=py311h295aba0_0_cpu
fastapi=0.103.0=pyhd8ed1ab_0
ffmpeg=4.2.2=h04105a8_0
filelock=3.9.0=py311hca03da5_0
freetype=2.12.1=h1192e45_0
frozenlist=1.3.3=py311h80987f9_0
fsspec=2023.4.0=py311hca03da5_0
gettext=0.21.0=h13f89a0_1
gflags=2.2.2=hc377ac9_0
giflib=5.2.1=h80987f9_3
glog=0.5.0=hc377ac9_0
gmp=6.2.1=hc377ac9_3
gmpy2=2.1.2=py311h40f64dc_0
gnutls=3.6.15=h887c41c_0
google-api-core=2.11.1=pyhd8ed1ab_0
google-auth=2.23.0=pyh1a96a4e_1
googleapis-common-protos=1.60.0=pyhd8ed1ab_0
greenlet=2.0.1=py311h313beb8_0
grpc-cpp=1.48.2=hc60591f_1
h11=0.14.0=pyhd8ed1ab_0
hnswlib=0.7.0=py311hddbb800_1
huggingface_hub=0.17.1=py_0
humanfriendly=10.0=py311h267d04e_4
icu=73.1=h313beb8_0
idna=3.4=py311hca03da5_0
importlib-metadata=6.0.0=py311hca03da5_0
ipykernel=6.25.2=pyh1050b4e_0
ipython=8.15.0=pyh31c8845_0
ipywidgets=8.1.1=pyhd8ed1ab_0
jedi=0.19.0=pyhd8ed1ab_0
jinja2=3.1.2=py311hca03da5_0
joblib=1.3.2=pypi_0
jpeg=9e=h80987f9_1
jupyter_client=7.4.9=pyhd8ed1ab_0
jupyter_core=5.3.0=py311hca03da5_0
jupyterlab_widgets=3.0.9=pyhd8ed1ab_0
krb5=1.20.1=hf3e1bf2_1
lame=3.100=h1a28f6b_0
langchain=0.0.292=pyhd8ed1ab_0
langsmith=0.0.37=pyhd8ed1ab_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libblas=3.9.0=16_osxarm64_openblas
libboost=1.82.0=hdbf59fe_1
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libcurl=8.2.1=h3e2b118_0
libcxx=16.0.6=h4653b0c_0
libdeflate=1.17=h80987f9_0
libedit=3.1.20221030=h80987f9_0
libev=4.33=h1a28f6b_1
libevent=2.1.12=h02f6b3c_1
libexpat=2.5.0=hb7217d7_1
libfaiss=1.7.4=h1234567_0_cpu
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h1a28f6b_2
libidn2=2.3.4=h80987f9_0
libjpeg-turbo=2.0.0=h1a28f6b_0
liblapack=3.9.0=16_osxarm64_openblas
libnghttp2=1.52.0=h62f6fdd_1
libopenblas=0.3.21=h269037a_0
libopus=1.3=h1a28f6b_1
libpng=1.6.39=h80987f9_0
libprotobuf=3.20.3=h514c7bf_0
libsodium=1.0.18=h27ca646_1
libsqlite=3.43.0=hb31c410_0
libssh2=1.10.0=h02f6b3c_2
libtasn1=4.19.0=h80987f9_0
libthrift=0.15.0=h73c2103_2
libtiff=4.5.1=h313beb8_0
libunistring=0.9.10=h1a28f6b_0
libvpx=1.10.0=hc377ac9_0
libwebp=1.2.4=ha3663a8_1
libwebp-base=1.2.4=h80987f9_1
libxml2=2.10.4=h0dcf63f_1
libzlib=1.2.13=h53f4e23_5
llvm-openmp=16.0.6=h1c12783_0
lz4=4.3.2=py311h854b99e_0
lz4-c=1.9.4=h313beb8_0
markupsafe=2.1.1=py311h80987f9_0
marshmallow=3.20.1=pyhd8ed1ab_0
marshmallow-enum=1.5.1=pyh9f0ad1d_3
matplotlib-inline=0.1.6=pyhd8ed1ab_0
monotonic=1.5=py_0
mpc=1.1.0=h8c48613_1
mpfr=4.0.2=h695f6f0_1
mpmath=1.3.0=py311hca03da5_0
multidict=6.0.2=py311h80987f9_0
multiprocess=0.70.15=py311hca03da5_0
mypy_extensions=1.0.0=pyha770c72_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.7.3=h84b5d62_1
networkx=2.8.8=pyhd8ed1ab_0
nltk=3.8.1=pypi_0
numexpr=2.8.4=py311h6dc990b_1
numpy=1.25.2=py311he598dae_0
numpy-base=1.25.2=py311hfbfe69c_0
onnxruntime=1.15.1=py311h17ae30d_3_cpu
openapi-schema-pydantic=1.2.4=pyhd8ed1ab_0
openh264=1.8.0=h98b2900_0
openssl=3.1.2=h53f4e23_0
orc=1.7.4=hdca1487_1
overrides=7.4.0=pyhd8ed1ab_0
packaging=23.1=py311hca03da5_0
pandas=2.0.3=py311h7aedaa7_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py311h313beb8_0
pip=23.2.1=py311hca03da5_0
platformdirs=3.10.0=pyhd8ed1ab_0
posthog=3.0.2=pyhd8ed1ab_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
protobuf=3.20.3=py311h313beb8_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=11.0.0=py311h7575258_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.8=py311h80987f9_0
pygments=2.16.1=pyhd8ed1ab_0
pymupdf=1.23.3=pypi_0
pymupdfb=1.23.3=pypi_0
pyopenssl=23.2.0=py311hca03da5_0
pysocks=1.7.1=py311hca03da5_0
python=3.11.5=h47c9636_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=1.0.0=pyhd8ed1ab_1
python-duckdb=0.8.1=py311h9ded451_0
python-flatbuffers=23.5.26=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=2.0.2=py311h80987f9_1
python_abi=3.11=3_cp311
pytorch=2.2.0.dev20230917=py3.11_0
pytz=2022.7=py311hca03da5_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=py311h80987f9_1
pyzmq=25.1.0=py311h313beb8_0
re2=2022.04.01=hc377ac9_0
readline=8.2=h1a28f6b_0
regex=2022.7.9=py311h80987f9_0
requests=2.31.0=py311hca03da5_0
rsa=4.9=pyhd8ed1ab_0
safetensors=0.3.2=py311hf0e4da2_0
scikit-learn=1.3.0=pypi_0
scipy=1.11.2=pypi_0
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=68.0.0=py311hca03da5_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=hc377ac9_0
sniffio=1.3.0=pyhd8ed1ab_0
sqlalchemy=1.4.39=py311h80987f9_0
sqlite=3.41.2=h80987f9_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.27.0=pyhd8ed1ab_0
stringcase=1.2.0=py_0
sympy=1.12=pyh04b8f61_3
tenacity=8.2.3=pyhd8ed1ab_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=hb8d0fd4_0
tokenizers=0.13.2=py311h3dd52b7_1
torchaudio=2.2.0.dev20230917=py311_cpu
torchvision=0.17.0.dev20230917=py311_cpu
tornado=6.3.2=py311h80987f9_0
tqdm=4.65.0=py311hb6e6a13_0
traitlets=5.10.0=pyhd8ed1ab_0
transformers=4.32.1=py311hca03da5_0
typing-extensions=4.7.1=py311hca03da5_0
typing_extensions=4.7.1=py311hca03da5_0
typing_inspect=0.9.0=pyhd8ed1ab_0
typing_utils=0.1.0=pyhd8ed1ab_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py311hca03da5_0
utf8proc=2.6.1=h1a28f6b_0
uvicorn=0.23.2=py311h267d04e_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.38.4=py311hca03da5_0
widgetsnbextension=4.0.9=pyhd8ed1ab_0
x264=1!152.20180806=h1a28f6b_0
xxhash=0.8.0=h1a28f6b_3
xz=5.4.2=h80987f9_0
yaml=0.2.5=h1a28f6b_0
yarl=1.8.1=py311h80987f9_0
zeromq=4.3.4=hbdafb3b_1
zipp=3.11.0=py311hca03da5_0
zlib=1.2.13=h53f4e23_5
zstandard=0.21.0=py311hae4035a_0
zstd=1.5.5=hd90d995_0
